refactor: change InfoBannerTransparent name
