+++ src/modules/userCheckIn/info-banner-compact/InfoBannerCompact.stories.ts
