+++ src/modules/userCheckIn/info-banner-compact/InfoBannerCompact.vue
